builder: remove daemon dependency in ContainerCreate()
